Did way more things than I should have in 1 commit
I guess I deleted old content from the Rails app that I copied. And I
sort of twiddled with the first page and design, plus started on getting
a page to show a single blog at a time.

It would be nice if I would do some of the transitions in javascript
rather than as page loads. But I'm not gonna do that till I know if this
is actually even going to be anything.